+++ src/levels/types.ts
@@ -1,6 +1,6 @@
 import type { CharacterOptions } from "../characters/types";
 import type { INonMainCharacter } from "../characters/interfaces";
-import type { Direction } from "../types";
+import type { HorizontalDirection } from "../types";
 
 export type LevelCharacter = {
   Constructor: new (options?: CharacterOptions) => INonMainCharacter;
@@ -18,5 +18,5 @@ export type LevelConstructorOptions = {
   title: string;
   objectives?: Array<() => boolean>;
   spawnableObstaclesPer30Second?: LevelCharacter[];
-  currentDirection?: Direction;
+  currentDirection?: HorizontalDirection;
 };